+++ admin/class-purger.php
@@ -188,7 +188,7 @@ abstract class Purger {
 
 		$this->log( 'Function purge_post BEGIN ===' );
 
-		if ( 1 === (int) $nginx_helper_admin->options['purge_homepage_on_edit'] ) {
+		if ( 1 === (int) $nginx_helper_admin->options['purge_homepage_on_edit'] && NGINX_HOME_PURGE ) {
 			$this->_purge_homepage();
 		}
 
@@ -271,7 +271,7 @@ abstract class Purger {
 
 		}
 
-		if ( $_purge_archive ) {
+		if ( $_purge_archive && NGINX_ARCHIVE_PURGE ) {
 
 			$_post_type_archive_link = get_post_type_archive_link( $_post_type );
 
@@ -345,7 +345,7 @@ abstract class Purger {
 			}
 		}
 
-		if ( $_purge_custom_taxa ) {
+		if ( $_purge_custom_taxa && NGINX_ARCHIVE_PURGE ) {
 
 			$custom_taxonomies = get_taxonomies(
 				array(
@@ -641,7 +641,7 @@ abstract class Purger {
 			$this->log( '# # # # #' );
 			$this->log( 'Function purge_on_post_moved_to_trash ( post id ' . $post->ID . ' ) BEGIN ===' );
 
-			if ( 1 === (int) $nginx_helper_admin->options['purge_homepage_on_del'] ) {
+			if ( 1 === (int) $nginx_helper_admin->options['purge_homepage_on_del'] && NGINX_HOME_PURGE ) {
 				$this->_purge_homepage();
 			}
